- working on PieChart.kt

+++ app/src/main/java/com/arbelkilani/bingetv/data/entities/genre/GenreData.kt
@@ -20,6 +20,6 @@ data class GenreData(
     @ColumnInfo(name = "count")
     var count: Int = 0,
 
-    var percentage: Int = 0
+    var percentage: Float = 0f
 
 )
