+++ app/src/main/java/com/arbelkilani/bingetv/domain/entities/genre/GenreEntity.kt
@@ -8,5 +8,5 @@ data class GenreEntity(
     var id: Int = -1,
     var name: String = "",
     var count: Int = 0,
-    var percentage: Int = 0
+    var percentage: Float = 0f
 ) : Parcelable
